fix(maskedinput): remove InputElementProps extra callbacks
IF-1495
