+++ packages/react-ui/components/MaskedInput/__tests__/MaskedInput-test.tsx
@@ -106,4 +106,24 @@ describe('MaskedInput', () => {
 
     expect(input).toHaveValue(expectedValue);
   });
+
+  it('onValueChange do not fire on focus', () => {
+    const valueChangeEvent = jest.fn();
+    render(<MaskedInput mask="+7 (999) 999 99 99" onValueChange={valueChangeEvent} />);
+
+    const input = screen.getByRole('textbox');
+    input.focus();
+
+    expect(valueChangeEvent).not.toHaveBeenCalled();
+  });
+
+  it('masked input calls onUnexpectedInput', () => {
+    const handleUnexpectedInput = jest.fn();
+    render(<MaskedInput mask="+7 (999) 999 99 99" onUnexpectedInput={handleUnexpectedInput} />);
+
+    const input = screen.getByRole('textbox');
+    fireEvent.input(input, { target: { value: 'A' } });
+
+    expect(handleUnexpectedInput).not.toHaveBeenCalled();
+  });
 });
